Refactor project structure by consolidating backend files, adding Docker support, and updating application properties
