+++ simpleEcom/backend/.gitattributes
