+++ simpleEcom/backend/src/main/resources/application.properties
@@ -1,0 +1,11 @@
+spring.application.name=simpleEcom
+
+server.port=8080
+
+spring.datasource.url=jdbc:h2:mem:simpleEcom
+spring.datasource.driverClassName=org.h2.Driver
+
+spring.jpa.show-sql=true
+
+spring.datasource.username=induwara
+spring.datasource.password=1234
