--- simpleEcom/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.application.name=simpleEcom
-
-server.port=8080
